+++ src/test/java/org/folio/linked/data/e2e/resource/ResourceControllerUpdateInstanceIT.java
@@ -110,7 +110,7 @@ class ResourceControllerUpdateInstanceIT extends ITBase {
     // Assert that the request is successful
     mockMvc.perform(updateRequest)
       .andExpect(status().isOk())
-      .andExpect(jsonPath("$.resource['http://bibfra.me/vocab/lite/Instance'].profileId").value(1));
+      .andExpect(jsonPath("$.resource['http://bibfra.me/vocab/lite/Instance'].profileId").value(3));
   }
 
   @Test
@@ -188,7 +188,7 @@ class ResourceControllerUpdateInstanceIT extends ITBase {
       {
         "resource": {
           "http://bibfra.me/vocab/lite/Instance": {
-            "profileId":  1,
+            "profileId":  3,
             "http://bibfra.me/vocab/marc/title": [
                 {
                   "http://bibfra.me/vocab/marc/Title": {