+++ src/test/java/org/folio/linked/data/test/json/IdentifierFieldResponseDeserializer.java
@@ -1,6 +1,6 @@
 package org.folio.linked.data.test.json;
 
-import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_EAN;
+import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_IAN;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_ISBN;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_LCCN;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_LOCAL;
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import java.io.IOException;
 import java.util.Map;
-import org.folio.linked.data.domain.dto.EanFieldResponse;
+import org.folio.linked.data.domain.dto.IanFieldResponse;
 import org.folio.linked.data.domain.dto.IdentifierFieldResponse;
 import org.folio.linked.data.domain.dto.IsbnFieldResponse;
 import org.folio.linked.data.domain.dto.LccnFieldResponse;
@@ -25,7 +25,7 @@ public class IdentifierFieldResponseDeserializer extends JsonDeserializer<Identi
   private static final Map<String, Class<? extends IdentifierFieldResponse>> IDENDTITY_MAP = Map.of(
     ID_LCCN.getUri(), LccnFieldResponse.class,
     ID_ISBN.getUri(), IsbnFieldResponse.class,
-    ID_EAN.getUri(), EanFieldResponse.class,
+    ID_IAN.getUri(), IanFieldResponse.class,
     ID_LOCAL.getUri(), LocalIdFieldResponse.class,
     ID_UNKNOWN.getUri(), OtherIdFieldResponse.class
   );